test: split packaged installation service tests
